requirements.txt
fastapi
uvicorn
passlib[bcrypt]
pyjwt
pymongo
mysql-connector-python
redis
